Extract DPM exception summary boundary
Evidence:

- python -m ruff format --check src/app/services/dpm_command_center_service.py src/app/services/dpm_command_center_exception_summary.py src/app/services/dpm_command_center_errors.py src/app/services/dpm_pm_operating_quality_service.py tests/unit/test_service_layer_boundaries.py: passed

- python -m ruff check src/app/services/dpm_command_center_service.py src/app/services/dpm_command_center_exception_summary.py src/app/services/dpm_command_center_errors.py src/app/services/dpm_pm_operating_quality_service.py tests/unit/test_service_layer_boundaries.py: passed

- python -m mypy src/app/services/dpm_command_center_service.py src/app/services/dpm_command_center_exception_summary.py src/app/services/dpm_command_center_errors.py src/app/services/dpm_pm_operating_quality_service.py: passed

- python -m pytest tests/unit/test_service_layer_boundaries.py tests/unit/test_dpm_command_center_service.py tests/integration/test_dpm_command_center_router.py -q: 74 passed
